fix inability to read some AV1 streams with RTSP (#6001)
Since v1.16.0, temporal unit delimiters were not stripped anymore from
AV1 streams. This has been restored, healing AV1 streams read with
RTSP.
