+++ internal/stream/rtp_decoder_test.go
@@ -1,382 +1,16 @@
-package stream
+package stream_test
 
 import (
 	"testing"
 
 	"github.com/bluenviron/gortsplib/v5/pkg/description"
 	"github.com/bluenviron/gortsplib/v5/pkg/format"
-	"github.com/bluenviron/mediamtx/internal/logger"
+	"github.com/bluenviron/mediamtx/internal/stream"
+	"github.com/bluenviron/mediamtx/internal/test"
 	"github.com/bluenviron/mediamtx/internal/unit"
 	"github.com/pion/rtp"
 	"github.com/stretchr/testify/require"
 )
-
-type nilLogger struct{}
-
-func (nilLogger) Log(logger.Level, string, ...any) {
-}
-
-func TestStream(t *testing.T) {
-	desc := &description.Session{Medias: []*description.Media{
-		{
-			Type:    description.MediaTypeVideo,
-			Formats: []format.Format{&format.H264{PacketizationMode: 1}},
-		},
-		{
-			Type:    description.MediaTypeVideo,
-			Formats: []format.Format{&format.VP8{}},
-		},
-	}}
-
-	strm := &Stream{
-		OrigDesc:          desc,
-		WriteQueueSize:    512,
-		RTPMaxPayloadSize: 1450,
-	}
-	err := strm.Initialize()
-	require.NoError(t, err)
-	defer strm.Close()
-
-	subStream := &SubStream{
-		Stream:        strm,
-		UseRTPPackets: false,
-	}
-	err = subStream.Initialize()
-	require.NoError(t, err)
-
-	r := &Reader{}
-
-	recv := make(chan struct{})
-
-	r.OnData(desc.Medias[0], desc.Medias[0].Formats[0], func(_ *unit.Unit) error {
-		close(recv)
-		return nil
-	})
-
-	strm.AddReader(r)
-	defer strm.RemoveReader(r)
-
-	subStream.WriteUnit(desc.Medias[0], desc.Medias[0].Formats[0], &unit.Unit{
-		PTS: 30000 * 2,
-		Payload: unit.PayloadH264{
-			{5, 2}, // IDR
-		},
-	})
-
-	<-recv
-
-	require.Equal(t, uint64(14), strm.InboundBytes())
-	require.Equal(t, uint64(14), strm.OutboundBytes())
-}
-
-func TestStreamSkipOutboundBytes(t *testing.T) {
-	desc := &description.Session{Medias: []*description.Media{
-		{
-			Type:    description.MediaTypeVideo,
-			Formats: []format.Format{&format.H264{PacketizationMode: 1}},
-		},
-		{
-			Type:    description.MediaTypeVideo,
-			Formats: []format.Format{&format.VP8{}},
-		},
-	}}
-
-	strm := &Stream{
-		OrigDesc:          desc,
-		WriteQueueSize:    512,
-		RTPMaxPayloadSize: 1450,
-	}
-	err := strm.Initialize()
-	require.NoError(t, err)
-	defer strm.Close()
-
-	subStream := &SubStream{
-		Stream:        strm,
-		UseRTPPackets: false,
-	}
-	err = subStream.Initialize()
-	require.NoError(t, err)
-
-	r := &Reader{
-		SkipOutboundBytes: true,
-	}
-
-	recv := make(chan struct{})
-
-	r.OnData(desc.Medias[0], desc.Medias[0].Formats[0], func(_ *unit.Unit) error {
-		close(recv)
-		return nil
-	})
-
-	strm.AddReader(r)
-	defer strm.RemoveReader(r)
-
-	subStream.WriteUnit(desc.Medias[0], desc.Medias[0].Formats[0], &unit.Unit{
-		PTS: 30000 * 2,
-		Payload: unit.PayloadH264{
-			{5, 2}, // IDR
-		},
-	})
-
-	<-recv
-
-	require.Equal(t, uint64(14), strm.InboundBytes())
-	require.Equal(t, uint64(0), strm.OutboundBytes())
-}
-
-func TestStreamResizeOversizedRTPPackets(t *testing.T) {
-	desc := &description.Session{Medias: []*description.Media{
-		{
-			Type: description.MediaTypeVideo,
-			Formats: []format.Format{&format.H264{
-				PacketizationMode: 1,
-				SPS: []byte{ // 1920x1080 baseline
-					0x67, 0x42, 0xc0, 0x28, 0xd9, 0x00, 0x78, 0x02,
-					0x27, 0xe5, 0x84, 0x00, 0x00, 0x03, 0x00, 0x04,
-					0x00, 0x00, 0x03, 0x00, 0xf0, 0x3c, 0x60, 0xc9, 0x20,
-				},
-				PPS: []byte{0x08, 0x06, 0x07, 0x08},
-			}},
-		},
-	}}
-
-	strm := &Stream{
-		OrigDesc:          desc,
-		WriteQueueSize:    512,
-		RTPMaxPayloadSize: 400,
-		Parent:            &nilLogger{},
-	}
-	err := strm.Initialize()
-	require.NoError(t, err)
-	defer strm.Close()
-
-	subStream := &SubStream{
-		Stream:        strm,
-		UseRTPPackets: true,
-	}
-	err = subStream.Initialize()
-	require.NoError(t, err)
-
-	r := &Reader{}
-
-	recv := make(chan *unit.Unit)
-	n := 0
-
-	r.OnData(desc.Medias[0], desc.Medias[0].Formats[0], func(u *unit.Unit) error {
-		switch n {
-		case 0:
-		case 1:
-			recv <- u
-		default:
-			t.Error("should not happen")
-		}
-		n++
-		return nil
-	})
-
-	strm.AddReader(r)
-	defer strm.RemoveReader(r)
-
-	subStream.WriteUnit(desc.Medias[0], desc.Medias[0].Formats[0], &unit.Unit{
-		PTS: 90000,
-		RTPPackets: []*rtp.Packet{
-			{
-				Header: rtp.Header{
-					Version:        2,
-					Marker:         true,
-					PayloadType:    96,
-					SequenceNumber: 122,
-					Timestamp:      45343,
-					SSRC:           563423,
-				},
-				Payload: []byte{1, 2, 3, 4},
-			},
-		},
-	})
-
-	oversizedPayload := make([]byte, 1000)
-	for i := range oversizedPayload {
-		oversizedPayload[i] = byte(i % 256)
-	}
-
-	subStream.WriteUnit(desc.Medias[0], desc.Medias[0].Formats[0], &unit.Unit{
-		PTS: 90000,
-		RTPPackets: []*rtp.Packet{
-			{
-				Header: rtp.Header{
-					Version:        2,
-					Marker:         true,
-					PayloadType:    96,
-					SequenceNumber: 123,
-					Timestamp:      45343,
-					SSRC:           563423,
-				},
-				Payload: oversizedPayload,
-			},
-		},
-	})
-
-	received := <-recv
-
-	require.Equal(t, 3, len(received.RTPPackets))
-
-	for i, pkt := range received.RTPPackets {
-		require.Equal(t, 123+uint16(i), pkt.SequenceNumber)
-		require.Equal(t, uint32(45343), pkt.Timestamp)
-	}
-
-	totalPayloadSize := 0
-	for _, pkt := range received.RTPPackets {
-		require.LessOrEqual(t, len(pkt.Payload), 400)
-		totalPayloadSize += len(pkt.Payload)
-	}
-
-	require.Equal(t, 1005, totalPayloadSize)
-}
-
-func TestStreamUpdateFormatParams(t *testing.T) {
-	for _, ca := range []string{"h264", "h265", "mpeg4video"} {
-		t.Run(ca, func(t *testing.T) {
-			var forma format.Format
-			var u *unit.Unit
-
-			switch ca {
-			case "h264":
-				sps := []byte{
-					0x67, 0x64, 0x00, 0x20, 0xac, 0xd9, 0x40, 0x78,
-					0x02, 0x27, 0xe5, 0x9a, 0x80, 0x80, 0x80, 0xa0,
-				}
-				pps := []byte{0x08, 0x07, 0x08, 0x09}
-
-				forma = &format.H264{
-					PacketizationMode: 1,
-					SPS:               []byte{0x67, 0x42, 0xc0, 0x28},
-					PPS:               []byte{0x08, 0x06},
-				}
-
-				u = &unit.Unit{
-					PTS: 90000,
-					Payload: unit.PayloadH264{
-						sps,    // New SPS
-						pps,    // New PPS
-						{5, 1}, // IDR
-					},
-				}
-
-			case "h265":
-				vps := []byte{0x40, 0x01, 0x0c, 0x01, 0xff, 0xfe}
-				sps := []byte{0x42, 0x01, 0x01, 0x01, 0x60, 0x00, 0x00, 0x04}
-				pps := []byte{0x44, 0x01, 0xc1, 0x73, 0xd1, 0x8a}
-
-				forma = &format.H265{
-					VPS: []byte{0x40, 0x01, 0x0c},
-					SPS: []byte{0x42, 0x01, 0x01},
-					PPS: []byte{0x44, 0x01, 0xc1},
-				}
-
-				u = &unit.Unit{
-					PTS: 90000,
-					Payload: unit.PayloadH265{
-						vps,                      // New VPS
-						sps,                      // New SPS
-						pps,                      // New PPS
-						{0x26, 0x01, 0x01, 0x02}, // IDR
-					},
-				}
-
-			case "mpeg4video":
-				config := []byte{
-					0x00, 0x00, 0x01, 0xb0, // Visual Object Sequence Start
-					0x02, 0x00, 0x00, 0x01, 0xb5, 0x8a,
-					0x14, 0x00, 0x00, 0x01, 0x00,
-				}
-
-				forma = &format.MPEG4Video{
-					Config: []byte{0x00, 0x00, 0x01, 0xb0, 0x01},
-				}
-
-				frame := make([]byte, 0, len(config)+20)
-				frame = append(frame, config...)
-				frame = append(frame, []byte{0x00, 0x00, 0x01, 0xb3}...) // Group of VOP
-				frame = append(frame, []byte{0x01, 0x02, 0x03, 0x04}...)
-
-				u = &unit.Unit{
-					PTS:     90000,
-					Payload: unit.PayloadMPEG4Video(frame),
-				}
-			}
-
-			media := &description.Media{
-				Type:    description.MediaTypeVideo,
-				Formats: []format.Format{forma},
-			}
-
-			desc := &description.Session{Medias: []*description.Media{media}}
-
-			strm := &Stream{
-				OrigDesc:          desc,
-				WriteQueueSize:    512,
-				RTPMaxPayloadSize: 1450,
-			}
-			err := strm.Initialize()
-			require.NoError(t, err)
-			defer strm.Close()
-
-			subStream := &SubStream{
-				Stream:        strm,
-				UseRTPPackets: false,
-			}
-			err = subStream.Initialize()
-			require.NoError(t, err)
-
-			r := &Reader{}
-			recv := make(chan struct{})
-
-			r.OnData(media, forma, func(_ *unit.Unit) error {
-				close(recv)
-				return nil
-			})
-
-			strm.AddReader(r)
-			defer strm.RemoveReader(r)
-
-			subStream.WriteUnit(media, forma, u)
-
-			<-recv
-
-			outDesc := strm.OutDescCopy()
-
-			// Verify that format parameters were updated
-			switch ca {
-			case "h264":
-				require.Equal(t, &format.H264{
-					SPS: []byte{
-						0x67, 0x64, 0x00, 0x20, 0xac, 0xd9, 0x40, 0x78,
-						0x02, 0x27, 0xe5, 0x9a, 0x80, 0x80, 0x80, 0xa0,
-					},
-					PPS:               []byte{0x08, 0x07, 0x08, 0x09},
-					PacketizationMode: 1,
-				}, outDesc.Medias[0].Formats[0])
-
-			case "h265":
-				require.Equal(t, &format.H265{
-					VPS: []byte{0x40, 0x01, 0x0c, 0x01, 0xff, 0xfe},
-					SPS: []byte{0x42, 0x01, 0x01, 0x01, 0x60, 0x00, 0x00, 0x04},
-					PPS: []byte{0x44, 0x01, 0xc1, 0x73, 0xd1, 0x8a},
-				}, outDesc.Medias[0].Formats[0])
-
-			case "mpeg4video":
-				require.Equal(t, &format.MPEG4Video{
-					Config: []byte{
-						0x00, 0x00, 0x01, 0xb0,
-						0x02, 0x00, 0x00, 0x01, 0xb5, 0x8a,
-						0x14, 0x00, 0x00, 0x01, 0x00,
-					},
-				}, outDesc.Medias[0].Formats[0])
-			}
-		})
-	}
-}
 
 var casesDecodeEncode = []struct {
 	name    string
@@ -1208,31 +842,31 @@ var casesDecodeEncode = []struct {
 	},
 }
 
-func TestStreamDecode(t *testing.T) {
+func TestRTPDecoder(t *testing.T) {
 	for _, ca := range casesDecodeEncode {
 		t.Run(ca.name, func(t *testing.T) {
 			desc := &description.Session{Medias: []*description.Media{{
 				Formats: []format.Format{ca.format},
 			}}}
 
-			strm := &Stream{
+			strm := &stream.Stream{
 				OrigDesc:          desc,
 				WriteQueueSize:    512,
 				RTPMaxPayloadSize: 1450,
-				Parent:            &nilLogger{},
+				Parent:            test.NilLogger,
 			}
 			err := strm.Initialize()
 			require.NoError(t, err)
 			defer strm.Close()
 
-			subStream := &SubStream{
+			subStream := &stream.SubStream{
 				Stream:        strm,
 				UseRTPPackets: true,
 			}
 			err = subStream.Initialize()
 			require.NoError(t, err)
 
-			r := &Reader{}
+			r := &stream.Reader{}
 			recv := make(chan *unit.Unit)
 
 			r.OnData(desc.Medias[0], ca.format, func(u *unit.Unit) error {
@@ -1256,80 +890,4 @@ func TestStreamDecode(t *testing.T) {
 			require.Equal(t, ca.decoded, received.Payload)
 		})
 	}
-}
-
-func TestStreamEncode(t *testing.T) {
-	for _, ca := range casesDecodeEncode {
-		t.Run(ca.name, func(t *testing.T) {
-			desc := &description.Session{Medias: []*description.Media{{
-				Formats: []format.Format{ca.format},
-			}}}
-
-			strm := &Stream{
-				OrigDesc:          desc,
-				WriteQueueSize:    512,
-				RTPMaxPayloadSize: 1450,
-				Parent:            &nilLogger{},
-			}
-			err := strm.Initialize()
-			require.NoError(t, err)
-			defer strm.Close()
-
-			subStream := &SubStream{
-				Stream:        strm,
-				UseRTPPackets: false,
-			}
-			err = subStream.Initialize()
-			require.NoError(t, err)
-
-			r := &Reader{}
-			recv := make(chan struct{})
-
-			r.OnData(desc.Medias[0], ca.format, func(u *unit.Unit) error {
-				for i := range min(len(u.RTPPackets), len(ca.encoded)) {
-					u.RTPPackets[i].Timestamp = ca.encoded[i].Timestamp
-					u.RTPPackets[i].SequenceNumber = ca.encoded[i].SequenceNumber
-					u.RTPPackets[i].SSRC = ca.encoded[i].SSRC
-				}
-				require.Equal(t, ca.encoded, u.RTPPackets)
-				close(recv)
-				return nil
-			})
-
-			strm.AddReader(r)
-			defer strm.RemoveReader(r)
-
-			subStream.WriteUnit(desc.Medias[0], ca.format, &unit.Unit{
-				Payload: ca.decoded,
-			})
-
-			<-recv
-		})
-	}
-}
-
-func TestStreamUpgradeH264PacketizationMode(t *testing.T) {
-	forma := &format.H264{
-		PacketizationMode: 0,
-		SPS:               []byte{0x67, 0x42, 0xc0, 0x28},
-		PPS:               []byte{0x08, 0x06},
-	}
-
-	strm := &Stream{
-		OrigDesc:          &description.Session{Medias: []*description.Media{{Formats: []format.Format{forma}}}},
-		WriteQueueSize:    512,
-		RTPMaxPayloadSize: 1450,
-		Parent:            &nilLogger{},
-	}
-	err := strm.Initialize()
-	require.NoError(t, err)
-	defer strm.Close()
-
-	outDesc := strm.OutDescCopy()
-
-	require.Equal(t, &format.H264{
-		PacketizationMode: 1,
-		SPS:               []byte{0x67, 0x42, 0xc0, 0x28},
-		PPS:               []byte{0x08, 0x06},
-	}, outDesc.Medias[0].Formats[0])
 }
