+++ internal/stream/stream_test.go
@@ -12,9 +12,260 @@ import (
 	"github.com/bluenviron/mediacommon/v2/pkg/formats/mp4/codecs"
 	"github.com/bluenviron/mediacommon/v2/pkg/formats/pmp4"
 	"github.com/bluenviron/mediamtx/internal/conf"
+	"github.com/bluenviron/mediamtx/internal/logger"
 	"github.com/bluenviron/mediamtx/internal/unit"
+	"github.com/pion/rtp"
 	"github.com/stretchr/testify/require"
 )
+
+type nilLogger struct{}
+
+func (nilLogger) Log(logger.Level, string, ...any) {
+}
+
+func TestStream(t *testing.T) {
+	desc := &description.Session{Medias: []*description.Media{
+		{
+			Type:    description.MediaTypeVideo,
+			Formats: []format.Format{&format.H264{PacketizationMode: 1}},
+		},
+		{
+			Type:    description.MediaTypeVideo,
+			Formats: []format.Format{&format.VP8{}},
+		},
+	}}
+
+	strm := &Stream{
+		OrigDesc:          desc,
+		WriteQueueSize:    512,
+		RTPMaxPayloadSize: 1450,
+	}
+	err := strm.Initialize()
+	require.NoError(t, err)
+	defer strm.Close()
+
+	subStream := &SubStream{
+		Stream:        strm,
+		UseRTPPackets: false,
+	}
+	err = subStream.Initialize()
+	require.NoError(t, err)
+
+	r := &Reader{}
+
+	recv := make(chan struct{})
+
+	r.OnData(desc.Medias[0], desc.Medias[0].Formats[0], func(_ *unit.Unit) error {
+		close(recv)
+		return nil
+	})
+
+	strm.AddReader(r)
+	defer strm.RemoveReader(r)
+
+	subStream.WriteUnit(desc.Medias[0], desc.Medias[0].Formats[0], &unit.Unit{
+		PTS: 30000 * 2,
+		Payload: unit.PayloadH264{
+			{5, 2}, // IDR
+		},
+	})
+
+	<-recv
+
+	require.Equal(t, uint64(14), strm.InboundBytes())
+	require.Equal(t, uint64(14), strm.OutboundBytes())
+}
+
+func TestStreamSkipOutboundBytes(t *testing.T) {
+	desc := &description.Session{Medias: []*description.Media{
+		{
+			Type:    description.MediaTypeVideo,
+			Formats: []format.Format{&format.H264{PacketizationMode: 1}},
+		},
+		{
+			Type:    description.MediaTypeVideo,
+			Formats: []format.Format{&format.VP8{}},
+		},
+	}}
+
+	strm := &Stream{
+		OrigDesc:          desc,
+		WriteQueueSize:    512,
+		RTPMaxPayloadSize: 1450,
+	}
+	err := strm.Initialize()
+	require.NoError(t, err)
+	defer strm.Close()
+
+	subStream := &SubStream{
+		Stream:        strm,
+		UseRTPPackets: false,
+	}
+	err = subStream.Initialize()
+	require.NoError(t, err)
+
+	r := &Reader{
+		SkipOutboundBytes: true,
+	}
+
+	recv := make(chan struct{})
+
+	r.OnData(desc.Medias[0], desc.Medias[0].Formats[0], func(_ *unit.Unit) error {
+		close(recv)
+		return nil
+	})
+
+	strm.AddReader(r)
+	defer strm.RemoveReader(r)
+
+	subStream.WriteUnit(desc.Medias[0], desc.Medias[0].Formats[0], &unit.Unit{
+		PTS: 30000 * 2,
+		Payload: unit.PayloadH264{
+			{5, 2}, // IDR
+		},
+	})
+
+	<-recv
+
+	require.Equal(t, uint64(14), strm.InboundBytes())
+	require.Equal(t, uint64(0), strm.OutboundBytes())
+}
+
+func TestStreamResizeOversizedRTPPackets(t *testing.T) {
+	desc := &description.Session{Medias: []*description.Media{
+		{
+			Type: description.MediaTypeVideo,
+			Formats: []format.Format{&format.H264{
+				PacketizationMode: 1,
+				SPS: []byte{ // 1920x1080 baseline
+					0x67, 0x42, 0xc0, 0x28, 0xd9, 0x00, 0x78, 0x02,
+					0x27, 0xe5, 0x84, 0x00, 0x00, 0x03, 0x00, 0x04,
+					0x00, 0x00, 0x03, 0x00, 0xf0, 0x3c, 0x60, 0xc9, 0x20,
+				},
+				PPS: []byte{0x08, 0x06, 0x07, 0x08},
+			}},
+		},
+	}}
+
+	strm := &Stream{
+		OrigDesc:          desc,
+		WriteQueueSize:    512,
+		RTPMaxPayloadSize: 400,
+		Parent:            &nilLogger{},
+	}
+	err := strm.Initialize()
+	require.NoError(t, err)
+	defer strm.Close()
+
+	subStream := &SubStream{
+		Stream:        strm,
+		UseRTPPackets: true,
+	}
+	err = subStream.Initialize()
+	require.NoError(t, err)
+
+	r := &Reader{}
+
+	recv := make(chan *unit.Unit)
+	n := 0
+
+	r.OnData(desc.Medias[0], desc.Medias[0].Formats[0], func(u *unit.Unit) error {
+		switch n {
+		case 0:
+		case 1:
+			recv <- u
+		default:
+			t.Error("should not happen")
+		}
+		n++
+		return nil
+	})
+
+	strm.AddReader(r)
+	defer strm.RemoveReader(r)
+
+	subStream.WriteUnit(desc.Medias[0], desc.Medias[0].Formats[0], &unit.Unit{
+		PTS: 90000,
+		RTPPackets: []*rtp.Packet{
+			{
+				Header: rtp.Header{
+					Version:        2,
+					Marker:         true,
+					PayloadType:    96,
+					SequenceNumber: 122,
+					Timestamp:      45343,
+					SSRC:           563423,
+				},
+				Payload: []byte{1, 2, 3, 4},
+			},
+		},
+	})
+
+	oversizedPayload := make([]byte, 1000)
+	for i := range oversizedPayload {
+		oversizedPayload[i] = byte(i % 256)
+	}
+
+	subStream.WriteUnit(desc.Medias[0], desc.Medias[0].Formats[0], &unit.Unit{
+		PTS: 90000,
+		RTPPackets: []*rtp.Packet{
+			{
+				Header: rtp.Header{
+					Version:        2,
+					Marker:         true,
+					PayloadType:    96,
+					SequenceNumber: 123,
+					Timestamp:      45343,
+					SSRC:           563423,
+				},
+				Payload: oversizedPayload,
+			},
+		},
+	})
+
+	received := <-recv
+
+	require.Equal(t, 3, len(received.RTPPackets))
+
+	for i, pkt := range received.RTPPackets {
+		require.Equal(t, 123+uint16(i), pkt.SequenceNumber)
+		require.Equal(t, uint32(45343), pkt.Timestamp)
+	}
+
+	totalPayloadSize := 0
+	for _, pkt := range received.RTPPackets {
+		require.LessOrEqual(t, len(pkt.Payload), 400)
+		totalPayloadSize += len(pkt.Payload)
+	}
+
+	require.Equal(t, 1005, totalPayloadSize)
+}
+
+func TestStreamUpgradeH264PacketizationMode(t *testing.T) {
+	forma := &format.H264{
+		PacketizationMode: 0,
+		SPS:               []byte{0x67, 0x42, 0xc0, 0x28},
+		PPS:               []byte{0x08, 0x06},
+	}
+
+	strm := &Stream{
+		OrigDesc:          &description.Session{Medias: []*description.Media{{Formats: []format.Format{forma}}}},
+		WriteQueueSize:    512,
+		RTPMaxPayloadSize: 1450,
+		Parent:            &nilLogger{},
+	}
+	err := strm.Initialize()
+	require.NoError(t, err)
+	defer strm.Close()
+
+	outDesc := strm.OutDescCopy()
+
+	require.Equal(t, &format.H264{
+		PacketizationMode: 1,
+		SPS:               []byte{0x67, 0x42, 0xc0, 0x28},
+		PPS:               []byte{0x08, 0x06},
+	}, outDesc.Medias[0].Formats[0])
+}
 
 func TestStreamAlwaysAvailableErrors(t *testing.T) {
 	for _, ca := range []struct {
